Reorganize files and add README
