+++ README.md
@@ -1,0 +1,12 @@
+# VectorKart
+
+A turn-based game based on [Racetrack](https://en.wikipedia.org/wiki/Racetrack_(game)) and made with WebGL.
+
+## Team: *Megatherium Within* 😎
+
+- Gabriel Cardozo (Product Owner)
+- Gobe Fang
+- Yesenia Puga
+- Calvin Owen
+- Brian Chung
+- Shauna Maloney
