+++ index.html
@@ -8,7 +8,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1">
 
         <!-- <link rel="apple-touch-icon" href="/apple-touch-icon.png"> -->
-        <link rel="stylesheet" href="style.css">
+        <link rel="stylesheet" href="/src/style.css">
         <!-- Place favicon.ico in the root directory -->
 
     </head>
@@ -23,8 +23,8 @@
 
         <canvas id="c"></canvas>
 
-        <script type="module" src="./index.js"></script>
-        <script type="module" src="./modules/webgl_utils.js"></script>
+        <script type="module" src="/src/index.js"></script>
+        <script type="module" src="/src/webgl_utils.js"></script>
 
     </body>
 </html>
