+++ resources/models/strange_shape.obj
