+++ resources/shaders/fragment_shader.glsl
