+++ resources/shaders/vertex_shader.glsl
