+++ src/vector.js
